+++ packages/@eventual/core-runtime/src/workflow.ts
@@ -1,7 +1,7 @@
 import { WorkflowContext } from "@eventual/core";
 
 declare module "@eventual/core" {
-  export interface Workflow<Input, Output> {
+  export interface Workflow<Name extends string, Input, Output> {
     definition: (
       input: Input,
       context: WorkflowContext