+++ packages/@eventual/core/src/entity/stream.ts
@@ -6,7 +6,7 @@ import {
 } from "../internal/service-spec.js";
 import type { ServiceContext } from "../service.js";
 import type { Entity, Attributes } from "./entity.js";
-import type { CompositeKeyPart, KeyMap } from "./key.js";
+import type { EntityCompositeKeyPart, KeyMap } from "./key.js";
 
 export interface EntityStreamContext {
   /**
@@ -17,9 +17,9 @@ export interface EntityStreamContext {
 
 export interface EntityStreamHandler<
   Attr extends Attributes = Attributes,
-  Partition extends CompositeKeyPart<Attr> = CompositeKeyPart<Attr>,
-  Sort extends CompositeKeyPart<Attr> | undefined =
-    | CompositeKeyPart<Attr>
+  Partition extends EntityCompositeKeyPart<Attr> = EntityCompositeKeyPart<Attr>,
+  Sort extends EntityCompositeKeyPart<Attr> | undefined =
+    | EntityCompositeKeyPart<Attr>
     | undefined
 > {
   /**
@@ -33,9 +33,9 @@ export interface EntityStreamHandler<
 
 export interface EntityStreamItemBase<
   Attr extends Attributes = Attributes,
-  Partition extends CompositeKeyPart<Attr> = CompositeKeyPart<Attr>,
-  Sort extends CompositeKeyPart<Attr> | undefined =
-    | CompositeKeyPart<Attr>
+  Partition extends EntityCompositeKeyPart<Attr> = EntityCompositeKeyPart<Attr>,
+  Sort extends EntityCompositeKeyPart<Attr> | undefined =
+    | EntityCompositeKeyPart<Attr>
     | undefined
 > {
   streamName: string;
@@ -45,9 +45,9 @@ export interface EntityStreamItemBase<
 
 export type EntityStreamItem<
   Attr extends Attributes = Attributes,
-  Partition extends CompositeKeyPart<Attr> = CompositeKeyPart<Attr>,
-  Sort extends CompositeKeyPart<Attr> | undefined =
-    | CompositeKeyPart<Attr>
+  Partition extends EntityCompositeKeyPart<Attr> = EntityCompositeKeyPart<Attr>,
+  Sort extends EntityCompositeKeyPart<Attr> | undefined =
+    | EntityCompositeKeyPart<Attr>
     | undefined
 > =
   | EntityStreamInsertItem<Attr, Partition, Sort>
@@ -56,9 +56,9 @@ export type EntityStreamItem<
 
 export interface EntityStreamInsertItem<
   Attr extends Attributes = Attributes,
-  Partition extends CompositeKeyPart<Attr> = CompositeKeyPart<Attr>,
-  Sort extends CompositeKeyPart<Attr> | undefined =
-    | CompositeKeyPart<Attr>
+  Partition extends EntityCompositeKeyPart<Attr> = EntityCompositeKeyPart<Attr>,
+  Sort extends EntityCompositeKeyPart<Attr> | undefined =
+    | EntityCompositeKeyPart<Attr>
     | undefined
 > extends EntityStreamItemBase<Attr, Partition, Sort> {
   newValue: Attr;
@@ -68,9 +68,9 @@ export interface EntityStreamInsertItem<
 
 export interface EntityStreamModifyItem<
   Attr extends Attributes = Attributes,
-  Partition extends CompositeKeyPart<Attr> = CompositeKeyPart<Attr>,
-  Sort extends CompositeKeyPart<Attr> | undefined =
-    | CompositeKeyPart<Attr>
+  Partition extends EntityCompositeKeyPart<Attr> = EntityCompositeKeyPart<Attr>,
+  Sort extends EntityCompositeKeyPart<Attr> | undefined =
+    | EntityCompositeKeyPart<Attr>
     | undefined
 > extends EntityStreamItemBase<Attr, Partition, Sort> {
   operation: "modify";
@@ -82,9 +82,9 @@ export interface EntityStreamModifyItem<
 
 export interface EntityStreamRemoveItem<
   Attr extends Attributes = Attributes,
-  Partition extends CompositeKeyPart<Attr> = CompositeKeyPart<Attr>,
-  Sort extends CompositeKeyPart<Attr> | undefined =
-    | CompositeKeyPart<Attr>
+  Partition extends EntityCompositeKeyPart<Attr> = EntityCompositeKeyPart<Attr>,
+  Sort extends EntityCompositeKeyPart<Attr> | undefined =
+    | EntityCompositeKeyPart<Attr>
     | undefined
 > extends EntityStreamItemBase<Attr, Partition, Sort> {
   operation: "remove";
@@ -93,10 +93,11 @@ export interface EntityStreamRemoveItem<
 }
 
 export interface EntityStream<
+  Name extends string,
   Attr extends Attributes,
-  Partition extends CompositeKeyPart<Attr>,
-  Sort extends CompositeKeyPart<Attr> | undefined
-> extends EntityStreamSpec<Attr, Partition, Sort> {
+  Partition extends EntityCompositeKeyPart<Attr>,
+  Sort extends EntityCompositeKeyPart<Attr> | undefined
+> extends EntityStreamSpec<Name, Attr, Partition, Sort> {
   kind: "EntityStream";
   handler: EntityStreamHandler<Attr, Partition, Sort>;
   sourceLocation?: SourceLocation;
@@ -104,31 +105,31 @@ export interface EntityStream<
 
 export function entityStream<
   Attr extends Attributes,
-  const Partition extends CompositeKeyPart<Attr>,
-  const Sort extends CompositeKeyPart<Attr> | undefined
+  const Partition extends EntityCompositeKeyPart<Attr>,
+  const Sort extends EntityCompositeKeyPart<Attr> | undefined
 >(
   ...args:
     | [
         name: string,
-        entity: Entity<Attr, Partition, Sort>,
+        entity: Entity<any, Attr, Partition, Sort>,
         handler: EntityStreamHandler<Attr, Partition, Sort>
       ]
     | [
         name: string,
-        entity: Entity<Attr, Partition, Sort>,
+        entity: Entity<any, Attr, Partition, Sort>,
         options: EntityStreamOptions<Attr, Partition, Sort>,
         handler: EntityStreamHandler<Attr, Partition, Sort>
       ]
     | [
         sourceLocation: SourceLocation,
         name: string,
-        entity: Entity<Attr, Partition, Sort>,
+        entity: Entity<any, Attr, Partition, Sort>,
         handler: EntityStreamHandler<Attr, Partition, Sort>
       ]
     | [
         sourceLocation: SourceLocation,
         name: string,
-        entity: Entity<Attr, Partition, Sort>,
+        entity: Entity<any, Attr, Partition, Sort>,
         options: EntityStreamOptions<Attr, Partition, Sort>,
         handler: EntityStreamHandler<Attr, Partition, Sort>
       ]
@@ -142,14 +143,14 @@ export function entityStream<
       ? [
           args[0],
           args[1] as string,
-          args[2] as Entity<Attr, Partition, Sort>,
+          args[2] as Entity<any, Attr, Partition, Sort>,
           ,
           args[3],
         ]
       : [
           ,
           args[0] as string,
-          args[1] as Entity<Attr, Partition, Sort>,
+          args[1] as Entity<any, Attr, Partition, Sort>,
           args[2] as EntityStreamOptions<Attr, Partition, Sort>,
           args[3],
         ];