+++ packages/@eventual/core/src/internal/open-api-spec.ts
@@ -90,9 +90,11 @@ export function generateOpenAPISpec(
             },
           },
           responses: {
-            default: {
-              content: { "application/json": { schema: command.output } },
-              description: `Default response for POST ${commandPath}`,
+            200: {
+              content: {
+                "application/json": { schema: command.output?.schema },
+              },
+              description: command.output?.description ?? "OK",
             } satisfies openapi.ResponseObject,
           },
         } satisfies openapi.OperationObject,
@@ -192,14 +194,18 @@ export function generateOpenAPISpec(
               : {}),
           },
         },
-        responses: {
-          default: {
-            description: `Default response for ${command.method} ${command.path}`,
-            content: {
-              "application/json": { schema: command.output },
-            },
-          },
-        },
+        responses: Object.fromEntries(
+          [
+            ...(command.output ? [command.output] : []),
+            ...(command.outputs ?? []),
+          ].map((o) => [
+            o.restStatusCode,
+            {
+              content: { "application/json": { schema: o.schema } },
+              description: o.description,
+            } satisfies openapi.ResponseObject,
+          ])
+        ),
       };
 
       return {